+++ Qracines/modules/tree_marking/create/tree_marking_create.py
@@ -43,12 +43,16 @@ class TreeMarkingCreateDialog(QDialog, FORM_CLASS):
 
         self.raster_controller = RasterController(
             ui=self,
-            raster_checkbox={
+             raster_checkbox={
                 #   'key':     'checkbox_name',
                 'r.seq.plt': 'cb_plt',
-                'r.alt.mnh': 'cb_mnh',
                 'r.ortho.irc': 'cb_irc',
                 'r.ortho.rgb': 'cb_rgb',
+                'r.alt.mnh.lidar': 'cb_mnh_lidar',
+                'r.alt.mnh.rge': 'cb_mnh_rge',
+                'r.alt.ombrage.mnh': 'cb_ombrage_mnh',
+                'r.alt.mnt.lidar': 'cb_mnt_lidar',
+                'r.alt.mnt.rge': 'cb_mnt_rge',
             })
 
         self.ess_selector = SpeciesSelector(