+++ Qracines/CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 ## [V1.0.10 - Unreleased] - 2026-06-08
 
 - Ajout de la couche ombrage dans expertise
-
+- Ajout de la couche ombrage dans tree_marking
 
 ## [V1.0.9] - 2026-05-04